Add dotnet-build-wrapper.sh
This script is an alternative to the build.sh provided by
upstream. The upstream build.sh does a lot of downloading
and processing of files that is overkill for my needs. This
script only builds the corehost and the NuGet packages, and
does not invoke the bespoke core-setup build system.